Misc cleanup in src/InfoBase
